+++ 35-seminar.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
