+++ 35-seminar.xcodeproj/xcshareddata/xcschemes/35-seminar0.xcscheme
