--- 35-seminar/35-seminar/Resources/Assets.xcassets/christmas_green.colorset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "extended-srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0.185",
-          "green" : "0.456",
-          "red" : "0.090"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "localizable" : true
-  }
-}
