+++ 35-seminar/Info.plist
