+++ 35-seminar/MainView.swift
@@ -1,0 +1,14 @@
+//
+//  MyView.swift
+//  35-seminar
+//
+//  Created by 김유림 on 10/5/24.
+//
+
+import UIKit
+import SnapKit
+
+class MyView: UIView {
+    
+
+}
