+++ 35-seminar/SceneDelegate.swift
@@ -1,21 +1,20 @@
 //
 //  SceneDelegate.swift
-//  AppstoreClone
+//  35-seminar
 //
-//  Created by 김유림 on 10/12/24.
+//  Created by 김유림 on 10/5/24.
 //
 
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
+    
     var window: UIWindow?
-
-
+    
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         self.window = UIWindow(windowScene: windowScene)
-        let navigationController = UINavigationController(rootViewController: DetailViewController())
+        let navigationController = UINavigationController(rootViewController: ViewController())
         self.window?.rootViewController = navigationController
         self.window?.makeKeyAndVisible()
     }
