--- [2주차 과제] 앱스토어/AppstoreClone/AppstoreClone/Presentation/Bases/BaseViewController.swift
@@ -1,21 +1,0 @@
-//
-//  BaseViewController.swift
-//  AppstoreClone
-//
-//  Created by 김유림 on 10/12/24.
-//
-
-import UIKit
-
-class BaseViewController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        // Do any additional setup after loading the view.
-    }
-    
-    func setHierarchy() { }
-    
-    func setConstraints() { }
-}
