--- [2주차 과제] 앱스토어/AppstoreClone/AppstoreClone/Presentation/Main/MainView.swift
@@ -1,12 +1,0 @@
-//
-//  MainView.swift
-//  AppstoreClone
-//
-//  Created by 김유림 on 10/12/24.
-//
-
-import UIKit
-
-class MainView: BaseView {
-
-}
