--- [2주차 과제] 앱스토어/AppstoreClone/AppstoreClone/Presentation/Main/MainViewController.swift
@@ -1,17 +1,0 @@
-//
-//  ViewController.swift
-//  AppstoreClone
-//
-//  Created by 김유림 on 10/12/24.
-//
-
-import UIKit
-
-class MainViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-}
-
