--- [2주차 과제] 앱스토어/AppstoreClone/AppstoreClone/Presentation/ReviewEdit/ReviewEditView.swift
@@ -1,12 +1,0 @@
-//
-//  ReviewEditView.swift
-//  AppstoreClone
-//
-//  Created by 김유림 on 10/12/24.
-//
-
-import UIKit
-
-class ReviewEditView: BaseView {
-
-}
