--- [2주차 과제] 앱스토어/AppstoreClone/AppstoreClone/Presentation/ReviewEdit/ReviewEditViewController.swift
@@ -1,29 +1,0 @@
-//
-//  ReviewEditViewController.swift
-//  AppstoreClone
-//
-//  Created by 김유림 on 10/12/24.
-//
-
-import UIKit
-
-class ReviewEditViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
