--- [2주차 과제] 앱스토어/AppstoreClone/AppstoreClone/Resources/Assets.xcassets/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
