--- [2주차 과제] 앱스토어/AppstoreClone/AppstoreClone/Resources/Assets.xcassets/toss_icon.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "toss_icon.png",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
